refactor(Icon): ♻️  Convert all usage to the new API [#3]
