+++ __mocks__/svgMock.js
@@ -1,0 +1,3 @@
+const React = require("react");
+
+module.exports = (props) => React.createElement("svg", props);
