+++ src/layouts/Footer/components/FooterSubscribeForm/FooterSubscribeForm.tsx
@@ -2,8 +2,7 @@ import { Form, Formik, FormikActions, FormikProps } from "formik";
 import { T } from "lioness";
 import React from "react";
 import Button from "src/components/discours-ui-kit/Button";
-import Icon from "src/components/discours-ui-kit/Icon";
-import { submit } from "src/components/discours-ui-kit/Icon/icons";
+import Icon, { submit } from "src/components/discours-ui-kit/Icon";
 import Input from "src/components/discours-ui-kit/Input";
 import Typography from "src/components/discours-ui-kit/Typography";
 import * as Yup from "yup";
@@ -69,7 +68,6 @@ class FooterSubscribeForm extends React.PureComponent<
     isSubmitting,
   }: FormikProps<FooterSubscribeFormValues>) => (
     <Form style={{ width: "100%" }}>
-      {/* TODO intl */}
       <Input
         type="email"
         name="email"
@@ -77,7 +75,7 @@ class FooterSubscribeForm extends React.PureComponent<
         placeholder="Ваш e-mail"
         append={
           <Button isLoading={isSubmitting} variant="button-regular" isInverse>
-            <Icon isInverse src={submit} />
+            <Icon isInverse component={submit} />
           </Button>
         }
       />
